refactor: move MemoryStream to arrow_deps
