+++ arrow_deps/Cargo.toml
@@ -13,6 +13,8 @@ arrow-flight = { git = "https://github.com/apache/arrow-rs", rev = "ed00e4d4a160
 # Turn off optional datafusion features (function packages)
 datafusion = { git = "https://github.com/apache/arrow-datafusion.git", rev = "f43dc444eb510f0ff170adecb5a23afb39c489a8", default-features = false }
 
+futures = "0.3"
+
 # Turn off the "arrow" feature; it currently has a bug that causes the crate to rebuild every time
 # and we're not currently using it anyway
 parquet = { git = "https://github.com/apache/arrow-rs", rev = "ed00e4d4a160cd5182bfafb81fee2240ec005014", default-features = false, features = ["snap", "brotli", "flate2", "lz4", "zstd"] }
