+++ server/src/db/streams.rs
@@ -44,42 +44,3 @@ impl futures::Stream for ReadFilterResultsStream {
 
     // TODO is there a useful size_hint to pass?
 }
-
-/// A RecordBatchStream created from a single RecordBatch
-///
-/// Unfortunately datafusion's MemoryStream is crate-local
-#[derive(Debug)]
-pub(crate) struct MemoryStream {
-    schema: SchemaRef,
-    batch: Option<RecordBatch>,
-}
-
-impl MemoryStream {
-    pub fn new(batch: RecordBatch) -> Self {
-        Self {
-            schema: batch.schema(),
-            batch: Some(batch),
-        }
-    }
-}
-
-impl RecordBatchStream for MemoryStream {
-    fn schema(&self) -> SchemaRef {
-        Arc::clone(&self.schema)
-    }
-}
-
-impl futures::Stream for MemoryStream {
-    type Item = ArrowResult<RecordBatch>;
-
-    fn poll_next(
-        mut self: std::pin::Pin<&mut Self>,
-        _: &mut Context<'_>,
-    ) -> Poll<Option<Self::Item>> {
-        Poll::Ready(self.batch.take().map(Ok))
-    }
-
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        (1, Some(1))
-    }
-}
